requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.1
dj-database-url==0.5.0
Django==3.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-forms-test==0.1.post1
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
